Split search bar into component
