+++ lib/view/screens/search.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:jisho_study_tool/bloc/search/search_bloc.dart';
-import 'package:jisho_study_tool/view/components/search/LanguageSelector.dart';
+import 'package:jisho_study_tool/view/components/search/search_bar.dart';
 import 'package:jisho_study_tool/view/screens/loading.dart';
 import 'package:jisho_study_tool/view/components/search/search_result_page/search_card.dart';
 
@@ -43,34 +43,5 @@ class _InitialView extends StatelessWidget {
     return Column(mainAxisAlignment: MainAxisAlignment.center, children: [
       SearchBar(),
     ]);
-  }
-}
-
-
-class SearchBar extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.symmetric(horizontal: 20.0),
-      child: Column(
-        children: [
-          TextField(
-            onSubmitted: (text) => BlocProvider.of<SearchBloc>(context)
-                .add(GetSearchResults(text)),
-            controller: TextEditingController(),
-            decoration: InputDecoration(
-              labelText: 'Search',
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(10.0),
-              ),
-            ),
-          ),
-          SizedBox(
-            height: 10.0,
-          ),
-          LanguageSelector()
-        ],
-      ),
-    );
   }
 }
